+++ examples/main-react16/src/pages/vue2/vue2.js
@@ -48,7 +48,7 @@ function Vue2 () {
         onMounted={mounted}
         onUnmount={unmount}
         // shadowDOM
-        // destory
+        // destroy
         // inline
         // disableScopecss
         // disableSandbox