+++ examples/main-vue2/src/pages/vite.vue
@@ -6,7 +6,7 @@
       inline
       disableSandbox
     >
-    <!-- destory inline scopecss='false' -->
+    <!-- destroy inline scopecss='false' -->
     </micro-app>
   </div>
 </template>