+++ examples/main-vue2/src/pages/vue3.vue
@@ -6,7 +6,7 @@
       :data='data'
       baseRoute='/vue3'
     >
-    <!-- destory inline scopecss='false' -->
+    <!-- destroy inline scopecss='false' -->
     </micro-app>
   </div>
 </template>