+++ examples/main-vue3-vite/src/pages/angular11.vue
@@ -7,7 +7,7 @@
     :data='data'
     inline
   >
-    <!-- destory inline scopecss='false' -->
+    <!-- destroy inline scopecss='false' -->
   </micro-app>
 </div>
 </template>