+++ examples/main-vue3-vite/src/pages/react17.vue
@@ -5,7 +5,7 @@
     url='http://localhost:3002/'
     :data='data'
   >
-    <!-- destory inline -->
+    <!-- destroy inline -->
   </micro-app>
 </div>
 </template>