+++ examples/main-vue3-vite/src/pages/vue3.vue
@@ -7,7 +7,7 @@
     macro
     baseRoute='/vue3'
   >
-    <!-- destory inline scopecss='false' -->
+    <!-- destroy inline scopecss='false' -->
   </micro-app>
 </div>
 </template>