+++ src/__tests__/create_app.test.ts
@@ -148,14 +148,14 @@ describe('create_app', () => {
     })
   })
 
-  // 非沙箱环境的umd，在destory卸载时，注册在window的函数应该删除
-  test('render umd app with disablesandbox & destory', async () => {
+  // 非沙箱环境的umd，在destroy卸载时，注册在window的函数应该删除
+  test('render umd app with disablesandbox & destroy', async () => {
     const microAppElement6 = document.createElement('micro-app')
     microAppElement6.setAttribute('name', 'test-app6')
     microAppElement6.setAttribute('library', 'umd-app1') // 自定义umd名称
     microAppElement6.setAttribute('url', `http://127.0.0.1:${ports.create_app}/umd1`)
     microAppElement6.setAttribute('disablesandbox', 'true')
-    microAppElement6.setAttribute('destory', 'true')
+    microAppElement6.setAttribute('destroy', 'true')
 
     appCon.appendChild(microAppElement6)